+++ models/classes/RdfObjectMapper/Annotation/RdfResourceAttributeType.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
